initial vulnerability fixes

+++ src/main/com/tcs/utx/digiframe/controller/HistoryController.java
@@ -1,9 +1,5 @@
 package com.tcs.utx.digiframe.controller;
 
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.ObjectOutputStream;
-import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.HashMap;
@@ -65,7 +61,7 @@ public class HistoryController {
 	private BrandingDetailsService brandingService;
 
 	@RequestMapping(value = "history", method = RequestMethod.GET,produces = "application/json; charset=utf-8")
-	public ResponseEntity<List<Map<String, Object>>> getHistory(@RequestParam(required=false)String histId) {
+	public ResponseEntity<List<Map<String, Object>>> getHistory() {
 		List<Map<String, Object>> data = new ArrayList<>();
 		Map<String, Object> map = new HashMap<String, Object>();
 		try {
@@ -85,16 +81,6 @@ public class HistoryController {
 				map.put(ACCESS_DENIED, "{\"ACCESS_DENIED\":\"Access Denied.\"}");
 				data.add(map);
 				return new ResponseEntity<>(data, HttpStatus.FORBIDDEN);
-			}
-			
-			if(histId!=null) {
-				List<String> content = new ArrayList<>();
-				
-				File f = new File(System.getProperty("user.dir")+histId);
-				content = Files.readAllLines(f.toPath());
-				map.put("history",content);
-				data.add(map);
-				return new ResponseEntity<>(data, HttpStatus.OK);
 			}
 
 			data = this.HistoryService.getHistory();
@@ -178,6 +164,18 @@ public class HistoryController {
 	public ResponseEntity<String> deleteHistory(@PathVariable int srno) {
 		LOG.info("HistoryController | deleteHistory Begin");
 		try {
+			int emp_id = BrandingDetailsController.getUser();
+
+			boolean isGuest = brandingService.isUserGuest();
+			if (isGuest) {
+				return new ResponseEntity<>(ACCESS_DENIED, HttpStatus.FORBIDDEN);
+			}
+
+			if (!this.permissionService.isOperationPermissible(TEXT_BUGHUNTR, TEXT_ADMIN, "View", emp_id, 0, 0)) {
+				LOG.info("HistoryController | Access Denied in deleteHistory");
+				return new ResponseEntity<>(ERROR_MSG, HttpStatus.FORBIDDEN);
+			}
+
 			this.HistoryService.deleteHistory(srno);
 			LOG.info("HistoryService | deleteHistory Exit");
 		} catch (DataAccessException e) {
